feat(editor): wire inspector transform edits to undo/redo + enhanced status bar
- Inspector transform edits now tracked via pending_transform_edit
- When editing stops, TransformEntityCommand recorded for undo/redo
- Status bar shows undo/redo descriptions (撤销/重做)
- Status bar shows play state indicator (▶ 运行中 / ⏸ 暂停)
- Fix clippy warnings (collapsible_if, useless_format, unnecessary_cast)

+++ crates/engine-render/src/resource/material.rs
@@ -243,7 +243,7 @@ impl MaterialStore {
         let id = self.next_id;
         self.next_id += 1;
 
-        let offset = (id - 1) as u64 * MATERIAL_STRIDE;
+        let offset = (id - 1) * MATERIAL_STRIDE;
 
         // Write base_color (vec4) to base_color_buffer
         queue.write_buffer(
